Breakout PlatformBans endpoints to separate API controller
